kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/fast/dom/HTMLFormElement/htmlformelement-nameless-control-submit-assertion.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<form id='f' onsubmit="didSubmit=true;return false;">
  <input id='i' type="text" value="a"></input>
  <input id='s' type="submit" value="submit"></input>
</form>

<script>
description('This test verifies that forms with nameless controls do not assert.');
var didSubmit = false;
var form = document.getElementById('f');
var s = document.getElementById('s');
s.click();
shouldBe("didSubmit", "true");

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
